+++ test/geocoding.js
@@ -107,13 +107,13 @@ test('MapboxClient#geocodeForward', function(t) {
 
     var tester = { client: function(opts) {
       var params = opts.params;
-      t.equals(params.types, 'country,region', 'types option is set');
+      t.equals(params.types, 'country,region,poi.landmark', 'types option is set');
       opts.callback();
       return { entity: function() {} };
     }};
 
     client.geocodeForward.apply(tester, ['Paris', {
-      types: 'country,region'
+      types: 'country,region,poi.landmark'
     }, function(err) {
       t.ifError(err);
       t.end();
@@ -234,6 +234,25 @@ test('MapboxClient#geocodeForward', function(t) {
     }]);
   });
 
+  t.test('options.limit', function(t) {
+    var client = new MapboxClient(process.env.MapboxAccessToken);
+    t.ok(client);
+
+    var tester = { client: function(opts) {
+      var params = opts.params;
+      t.equals(params.limit, 1, 'limit option is set');
+      opts.callback();
+      return { entity: function() {} };
+    }};
+
+    client.geocodeForward.apply(tester, ['Paris', {
+      limit: 1
+    }, function(err) {
+      t.ifError(err);
+      t.end();
+    }]);
+  });
+
   t.end();
 });
 
@@ -280,6 +299,27 @@ test('MapboxClient#geocodeReverse', function(t) {
     client.geocodeReverse.apply(tester, [
     { latitude: 33.6875431, longitude: -95.4431142 },
     { types: 'country,region' },
+    function(err) {
+      t.ifError(err);
+      t.end();
+    }]);
+  });
+
+  t.test('options.limit', function(t) {
+    var client = new MapboxClient(process.env.MapboxAccessToken);
+    t.ok(client);
+
+    var tester = { client: function(opts) {
+      var params = opts.params;
+      t.equals(params.limit, 2, 'limit option is set');
+      t.equals(params.types, 'poi', 'single types option is set');
+      opts.callback();
+      return { entity: function() {} };
+    }};
+
+    client.geocodeReverse.apply(tester, [
+    { latitude: 33.6875431, longitude: -95.4431142 },
+    { types: 'poi', limit: 2 },
     function(err) {
       t.ifError(err);
       t.end();